Move Parser class into parser package and under planitarium package
